update rhine gpiomux layout
